progressao/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
dj-database-url==2.1.0
Django==5.0
django-bootstrap4==23.2
django-braces==1.15.0
django-cleanup==8.0.0
django-crispy-forms==2.1
django-heroku==0.3.1
idna==3.6
Pillow==10.1.0
psycopg2==2.9.9
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
whitenoise==6.6.0
gunicorn==21.2.0
